add server-queue
