+++ tools/server/server-task.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "common.h"
 #include "llama.h"
 #include "chat.h"
